initial restructure, wip
