+++ src/modules/post/ui/graphql/post_add.graphql
