+++ src/modules/post/ui/graphql/post_comment_delete.graphql
