+++ src/modules/post/ui/graphql/post_delete.graphql
