+++ src/modules/post/ui/graphql/post_get.graphql
