+++ src/modules/post/ui/graphql/post_subscription.graphql
