+++ src/modules/post/ui/graphql/posts_get.graphql
