+++ src/modules/post/ui/graphql/posts_subscription.graphql
