+++ src/server/api/resolvers.js
@@ -1,0 +1,6 @@
+import postResolvers from '../../modules/post/server/api/post_resolvers'
+
+export default [
+  postResolvers,
+  // add more resolvers here
+];
