--- src/server/api/resolvers/index.js
@@ -1,6 +1,0 @@
-import postResolvers from './post_resolvers'
-
-export default [
-  postResolvers,
-  // add more resolvers here
-];
