+++ src/server/api/schema_defs.js
@@ -1,5 +1,5 @@
 import schema from './schema_def.graphqls'
-import postSchema from './post_def.graphqls'
+import postSchema from '../../modules/post/server/api/post_schema_def.graphqls'
 
 export default [
   schema,
